requirements.txt
Flask
gunicorn
sqlalchemy
mysqlclient
python-dotenv